% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoB.R
\name{LoB}
\alias{LoB}
\title{Limit of Blank (LoB) Calculation}
\usage{
LoB(df, alpha = 0.05, parametric = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{alpha}{Alpha (type I error)}

\item{parametric}{Parametric (TRUE) or non-parametric (FALSE). Default is non-parametric.}
}
\description{
Limit of Blank (LoB) Calculation
}
