% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD.R
\name{LoD}
\alias{LoD}
\title{Limit of Detection (LoD) Calculation}
\usage{
LoD(
  df,
  col_values,
  beta = 0.05,
  LoB,
  approach = c("classical", "precision profile", "probit")
)
}
\arguments{
\item{df}{data frame for one reagent lot}

\item{col_values}{Name (in quotes) of the column with measurements.}

\item{beta}{type II error}

\item{LoB}{limit of blank}

\item{approach}{method of LoD calculation}
}
\description{
Limit of Detection (LoD) Calculation
}
