% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD_classical.R
\name{LoD_classical}
\alias{LoD_classical}
\title{Limit of Detection (LoD) Calculation--Classical Approach}
\usage{
LoD_classical(
  df,
  col_lot = NULL,
  col_sample,
  col_value,
  LoB,
  beta = 0.05,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with low-level samples. Must have at least one column with sample
and one column with measurement values. Column for reagent lot is optional.}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (default).
To split the LoD results by any other variable (e.g. lab), simply include the name
of this other variable here and set always_sep_lots = TRUE.}

\item{col_sample}{Name (in quotes) of any column with unique values that correspond to each
sample. Does not need to be numeric.}

\item{col_value}{Name (in quotes) of the column with measurements.}

\item{LoB}{Limit of blank (LoB). This value should be the same for all reagent lots.}

\item{beta}{Type II error. Default is 0.05.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns a list with the limit of detection (LoD) value(s) for each reagent lot
(if evaluated separately) and the overall reported LoD.
}
\description{
When should I use the classical approach to calculate LoD?
}
\details{
From CLSI EP17 guidelines:
This is the original approach from the previous version of this document, which has been
widely used for many chemistry and immunochemical measurement procedures. It uses
measurements made on...a set of low level samples containing a measurand targeted
around the assumed LoD... An underlying assumption is that variability of measurement
results is reasonably consistent across the low level samples.

CLSI EP17 Requirements:
\itemize{
\item Two reagent lots
\item One instrument system
\item Three days
\item Four samples
\item Two replicates per sample (for each reagent lot, day, and instrument system
combination)
\item 60 total replicates per reagent lot
}
}
\examples{
# CLSI EP17 Appendix A
Reagent <- c(rep(1, 12*5), rep(2, 12*5))
Day <- rep(rep(c(1, 2, 3), each = 4, times = 10))
Sample <- rep(c(1, 2, 3, 4, 5), each = 12, times = 2)
Replicate <- rep(c(1, 2, 3, 4), times = 3*5*2)
Measurement <- c(21.0, 22.8, 28.2, 25.9, 26.4, 28.3, 20.7, 21.9,
            24.7, 22.5, 28.5, 29.2, 13.3, 12.6, 18.2, 14.7,
            17.8, 14.0, 14.1, 12.5, 11.3, 12.2, 16.2, 13.9,
            12.8, 12.9, 17.4, 16.0, 15.9, 14.1, 11.3, 9.4,
            10.6, 13.6, 17.6, 14.9, 17.3, 19.2, 21.5, 22.2,
            24.1, 25.8, 16.0, 16.4, 24.9, 23.8, 22.1, 26.1,
            19.2, 22.7, 28.3, 26.2, 25.1, 30.3, 23.4, 19.2,
            26.3, 23.1, 27.5, 30.1, 22.0, 22.5, 21.8, 22.1,
            20.3, 21.0, 25.3, 26.0, 27.2, 25.1, 25.3, 25.3,
            15.6, 21.2, 14.8, 14.9, 16.0, 15.8, 21.6, 22.8,
            15.3, 18.7, 18.3, 19.5, 13.0, 15.9, 9.0, 7.0,
            13.4, 8.5, 16.3, 18.1, 12.4, 11.1, 11.3, 10.1,
            18.8, 17.6, 14.1, 14.9, 19.2, 15.8, 19.8, 21.4,
            18.0, 18.0, 19.6, 23.1, 32.9, 30.4, 29.4, 27.6,
            27.7, 30.6, 31.4, 30.4, 32.5, 28.9, 29.8, 35.1)

LoD_C_df <- data.frame(Reagent, Day, Sample, Replicate, Measurement)

results <- LoD_classical(df = LoD_C_df,
                         col_lot = "Reagent",
                         col_sample = "Sample",
                         col_value = "Measurement",
                         LoB = 7.5)

}
