% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD_precision_profile.R
\name{LoD_precision_profile}
\alias{LoD_precision_profile}
\title{Limit of Detection (LoD) Calculation--Precision Profile Approach}
\usage{
LoD_precision_profile(
  df,
  col_lot,
  col_value,
  LoB,
  approach = c("classical", "precision profile", "probit"),
  beta = 0.05,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with low-level samples.}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (no quotes).}

\item{col_value}{Name (in quotes) of the column with measurements.}

\item{LoB}{Limit of blank (LoB). This value should be the same for all reagent lots.}

\item{beta}{Type II error. Default is 0.05.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns the limit of detection (LoD) value as calculated with the precision
profile approach.
}
\description{
Limit of Detection (LoD) Calculation--Precision Profile Approach
}
