% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD_probit.R
\name{LoD_probit}
\alias{LoD_probit}
\title{Limit of Detection (LoD) Calculation--Probit Approach}
\usage{
LoD_probit(
  df,
  col_lot,
  col_conc,
  col_obs_pos,
  col_tot,
  LoB,
  beta = 0.05,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with...}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (no quotes).}

\item{col_conc}{Name (in quotes) of the column with the concentration.}

\item{col_obs_pos}{Name (in quotes) of the column with the number of positive results.}

\item{col_tot}{Name (in quotes) of the column with the total number of results.}

\item{LoB}{Limit of blank (LoB). Can be set to zero if it can be confirmed that LoB = 0, or
calculated using the classical approch.}

\item{beta}{Type II error. Default is 0.05.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\description{
Limit of Detection (LoD) Calculation--Probit Approach
}
