% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoQ_total_error.R
\name{LoQ_total_error}
\alias{LoQ_total_error}
\title{Limit of Quantitation (LoQ) Calculation--Total Error Approach}
\usage{
LoQ_total_error(
  df,
  col_lot = NULL,
  col_sample,
  col_ref,
  col_value,
  accuracy_goal,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with low-level samples.}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (default).}

\item{col_sample}{Name (in quotes) of any column with unique values that correspond to each
sample. Can be NULL (default) if n_samples is provided. Does not need to be numeric.}

\item{col_ref}{Name (in quotes) of the column with reference/true values.}

\item{col_value}{Name (in quotes) of the column with measurements.}

\item{accuracy_goal}{Desired maximum total error as a percentage between 0 and 100.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns a list with a table of the total errors for each sample in each lot,
a plot with these total errors by mean measurement, the LoQ values values for each reagent
lot if evaluated separately, and the reported overall LoQ.
}
\description{
Note: This function only fits a linear model to predict total error from mean measurement because
this model is not used in the actual calculation of LoQ; it simply provides some insight into the
range of measurement values that may be associated with a total error below the accuracy goal.
}
\examples{
# CLSI EP17 Appendix D2
Reagent = c(rep(1, 9*5), rep(2, 9*5))
Day = rep(rep(c(1, 2, 3), each = 3, times = 10))
Sample = rep(c(1, 2, 3, 4, 5), each = 9, times = 2)
Reference_Value = rep(c(38.2, 47.1, 44.7, 36.5, 42.8), each = 9, times = 2)
Replicate = rep(c(1, 2, 3), times = 3*5*2)
Measurement = c(36.7, 37.9, 38.3, 36.8, 33.5, 39.2, 41.3, 37.9, 34.9,
49.9, 50, 48.1, 47.8, 43.9, 45.6, 45.4, 51.5, 45.8,
46.1, 43.1, 39.4, 47.3, 45.8, 44.8, 44.6, 47.3, 38.9,
33.3, 34.2, 34.5, 43.1, 34, 37.1, 35.3, 32.4, 36, 42.9,
41.8, 43.8, 46.3, 43.3, 46, 42.6, 41.4, 42.8, 38.5, 41,
43.2, 36.8, 42.1, 35.8, 36.8, 44.1, 39.5, 45.8, 47.8,
46.6, 46.9, 51.3, 50.5, 44.3, 47.5, 52.4, 46.7, 43.6,
42.4, 46.5, 47.9, 42.7, 42.1, 43.4, 44.7, 35.5, 40, 34,
32.9, 33.1, 38.6, 36.2, 41.4, 33, 42, 44.1, 43.2, 46.6,
45.5, 43.5, 41.4, 48.2, 45.7)

LoQ_TE_df <- data.frame(Reagent, Day, Sample, Reference_Value, Replicate, Measurement)

results <- LoQ_total_error(df = LoQ_TE_df,
                           col_lot = "Reagent",
                           col_sample = "Sample",
                           col_ref = "Reference_Value",
                           col_value = "Measurement",
                           accuracy_goal = 21.6)

}
