% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Double input value}
\usage{
test_function(x)
}
\arguments{
\item{x}{Any value}
}
\value{
Returns double the input value.
}
\description{
Double input value
}
\examples{
val = 3
test_function(val)
}
