% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoB.R
\name{LoB}
\alias{LoB}
\title{Limit of Blank (LoB) Calculation}
\usage{
LoB(
  df,
  col_lot,
  col_sample,
  col_value,
  alpha = 0.05,
  parametric = FALSE,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can also be NULL (no quotes).}

\item{col_sample}{Name (in quotes) of the column with sample number.}

\item{col_value}{Name (in quotes) of the column with measurements.}

\item{alpha}{Alpha (type I error). Default is 0.05.}

\item{parametric}{Parametric (TRUE) or non-parametric (FALSE). Default is non-parametric.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns the limit of blank (LoB) value(s).
}
\description{
Limit of Blank (LoB) Calculation
}
\examples{
reagent_lot <- c(rep(1, 12*5), rep(2, 12*5))
day <- rep(rep(c(1, 2, 3), each = 4, times = 10))
sample <- rep(c(1, 2, 3, 4, 5), each = 12, times = 2)
replicate <- rep(c(1, 2, 3, 4), times = 3*5*2)
pg.ml <- c(2.6, -.8, 5.5, 6.0, 4.5, .6, -2.3, 3.4,
            5.9, 7.6, 4.1, -1.4, 1.0, 2.9, 4.9, 8.0,
            6.9, 5.0, 3.4, 1.2, 6.5, 5.6, -2.2, 2.3,
            -4.4, -3.4, 7.0, 6.9, 4.3, 3.2, -1.4, 4.2,
            5.9, 7.6, 3.8, 5.8, 1.5, -1.9, 5.1, 5.7,
            4.1, 4.5, -.6, .5, 5.4, 7.6, 4.4, 6.6,
            1.2, -.7, 6.1, 5.1, 4.8, 3.3, -2.8, -1.4,
            8.7, 3.6, 5.1, 3.5, 4.6, 4.1, 1.6, 3.7,
            2.2, .7, 4.6, 2.6, 1.1, -4.4, .9, .7,
            9.2, 8.3, 4.8, 5.4, 4.8, 6.3, 5.4, 9.6,
            7.7, 3.1, 6.1, 10.0, 6.1, 3.2, 3.9, 1.4,
            3.1, 4.1, 1.0, 3.4, .1, .4, 2.9, -1.6,
            4.0, 11.5, 4.5, 3.6, 4.4, 6.8, 7.1, 4.2,
            3.7, 3.7, 5.3, 4.5, 4.0, 6.2, -.2, 2.3,
            1.6, 2.6, 6.4, 5.7, 4.2, 3.7, 1.4, 1.5)
blank_df <- data.frame(reagent_lot, day, sample, replicate, pg.ml)

LoB_vals <- LoB(blank_df, col_lot = "reagent_lot", col_sample = "sample", col_value = "pg.ml")

}
