% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoB.R
\name{LoB}
\alias{LoB}
\title{Limit of Blank (LoB) Calculation}
\usage{
LoB(df, col_lot, col_sample, col_values, alpha = 0.05, parametric = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL.}

\item{col_sample}{Name (in quotes) of the column with sample number.}

\item{col_values}{Name (in quotes) of the column with measurements.}

\item{alpha}{Alpha (type I error). Default is 0.05.}

\item{parametric}{Parametric (TRUE) or non-parametric (FALSE). Default is non-parametric.}
}
\description{
Limit of Blank (LoB) Calculation
}
