% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD.R
\name{LoD}
\alias{LoD}
\title{Limit of Detection (LoD) Calculation}
\usage{
LoD(
  df,
  alpha = 0.05,
  LoB,
  approach = c("classical parametric", "classical nonparametric", "precision profile",
    "probit")
)
}
\arguments{
\item{df}{data frame for one reagent lot}

\item{alpha}{alpha level (type I error)}

\item{LoB}{limit of blank}

\item{approach}{}
}
\description{
Limit of Detection (LoD) Calculation
}
