% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD_precision_profile.R
\name{LoD_precision_profile}
\alias{LoD_precision_profile}
\title{Limit of Detection (LoD) Calculation--Precision Profile Approach}
\usage{
LoD_precision_profile(
  df,
  col_lot,
  col_sample,
  col_avg,
  col_sd_wl,
  LoB_val,
  N,
  model = c("lowest AIC", "linear", "quadratic", "sadler"),
  sadler_start,
  beta = 0.05,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with with the mean value and within-laboratory precision (obtained
from CLSI EP05) for each sample from each reagent lot.}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (no quotes).}

\item{col_sample}{Name (in quotes) of the column with sample number.}

\item{col_avg}{Name (in quotes) of the column with measurements.}

\item{col_sd_wl}{Name (in quotes) of the column with within-lab precision.}

\item{LoB_val}{Limit of blank (LoB).}

\item{N}{Total number of measurement values per sample per reagent lot (number of measurements
per sample times number of sample). If all reagent lots are pooled, use the total number
of measurement values per sample (sum across all lots).}

\item{model}{Choice of precision model. Default is to choose either the linear or quadratic model
based on AIC. One can also select to use a linear, quadratic, or Sadler model, the three models
most widely used in clinical literature (according to CLSI guidelines).}

\item{sadler_start}{Vector of length 3 with starting coefficient values for the Sadler model,
with the form (beta0 + beta1*x)^beta2.}

\item{beta}{Type II error. Default is 0.05.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns a list with the limit of detection (LoD) value as calculated with the precision
profile approach, precision model coefficients, and precision profile plot.
}
\description{
Limit of Detection (LoD) Calculation--Precision Profile Approach
}
\examples{
reagent_lot <- c(rep(1, 6), rep(2, 6))
sample <- rep(c("A","B","C","D","E","F"), times = 2)
avg <- c(.69, 1.42, 2.65, 4.08, 6.08, 10.36, .78, 1.73, 2.89, 3.82, 6.33, 10.92)
sd_wl <- c(.39, .39, .46, .55, .64, 1.12, .29, .54, .55, .63, .82, 1.38)

LoD_PP_df <- data.frame(reagent_lot, sample, avg, sd_wl)

LoD_precision_profile(df = LoD_PP_df, col_lot = "reagent_lot", col_sample = "sample",
col_avg = "avg", col_sd_wl = "sd_wl", LoB = 0.51, N = 80*6)

}
