% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoD_probit.R
\name{LoD_probit}
\alias{LoD_probit}
\title{Limit of Detection (LoD) Calculation--Probit Approach}
\usage{
LoD_probit(
  df,
  col_lot,
  col_conc,
  col_obs_pos,
  col_tot,
  LoB,
  log10_trans = FALSE,
  beta = 0.05,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with...}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (no quotes).}

\item{col_conc}{Name (in quotes) of the column with the concentration.}

\item{col_obs_pos}{Name (in quotes) of the column with the number of positive results.}

\item{col_tot}{Name (in quotes) of the column with the total number of results.}

\item{LoB}{Limit of blank (LoB). If the false positives in each reagent lot do not exceed
(100*alpha)\%, the LoB can be set to zero. Otherwise, calculate the LoB using the classical
approach.}

\item{log10_trans}{According to CLSI EP17 guidelines, a log 10 transformation of concentration
values can lead to a better probit fit. If FALSE, a note will indicate if the model(s) with a log
10 transformation of concentration values have a better fit.}

\item{beta}{Type II error. Default is 0.05.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns a list with the limit of detection (LoD) value as calculated with the probit
approach, probit model coefficients, probit model fitted values and confidence intervals, and
probit model plot.
}
\description{
Limit of Detection (LoD) Calculation--Probit Approach
}
\examples{
conc <- rep(c(0, .006, .014, .025, .05, .15, .3, .5), each = 3)
reagent_lot <- rep(c(1, 2, 3), times = 8)
obs_positive <- c(0, 0, 0, 11, 12, 22, 15, 22, 31, 23, 28, 27, 29,
32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32)
total_calls <- c(22, 22, 22,30, 30, 34, 30, 30, 34, rep(32, 15))

LoD_P_df <- data.frame(conc, reagent_lot, obs_positive, total_calls)

LoD_probit(df = LoD_P_df, col_lot = "reagent_lot", col_conc = "conc",
col_obs_pos = "obs_positive", col_tot = "total_calls", LoB = 0)

}
