% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoQ_functional.R
\name{LoQ_functional}
\alias{LoQ_functional}
\title{Limit of Quantitation (LoQ) Calculation with the Functional Sensitivity Approach}
\usage{
LoQ_functional(
  df,
  col_lot,
  col_sample,
  col_avg,
  col_sd_wl,
  target_cv,
  model = c("power", "power w intercept"),
  coeff_start,
  always_sep_lots = FALSE
)
}
\arguments{
\item{df}{A data frame with...}

\item{col_lot}{Name (in quotes) of the column with reagent lot number. Can be NULL (no quotes).}

\item{col_sample}{Name (in quotes) of the column with sample number.}

\item{col_avg}{Name (in quotes) of the column with measurements.}

\item{col_sd_wl}{Name (in quotes) of the column with within-lab precision.}

\item{target_cv}{Desired coefficient of variation (sd/mean)*100.}

\item{model}{Select the power function (default) or power function with intercept to model
coefficient of variation (CV) as a function of measurement value.}

\item{coeff_start}{Vector with starting coefficient values for the selected model. Should have
length 2 for the power function, or length 3 for the power function with intercept.}

\item{always_sep_lots}{If FALSE, reagent lots are evaluated according to CLSI guidelines
(all lots evaluated separately if 2 or 3 lots, and all lots evaluated together if >3 lots).
If TRUE, all reagent lots are evaluated separately regardless of the number of lots.
Default is FALSE.}
}
\value{
Returns the limit of quantitation (LoQ) value, model coefficients, and plot with fitted
values from the model.
}
\description{
Limit of Quantitation (LoQ) Calculation with the Functional Sensitivity Approach
}
\examples{
reagent_lot <- rep(c(1, 2), each = 9)
sample <- rep(1:9, times = 2)
avg <- c(.04, .053, .08, .111, .137, .164, .190, .214, .245,
.041, .047, .077, .106, .136, .159, .182, .205, .234)
sd_wl <- c(.016, .016, .016, .017, .014, .012, .011, .016, .013,
.018, .014, .012, .019, .016, .015, .015, .016, .014)

LoQ_F_df <- data.frame(reagent_lot, sample, avg, sd_wl)

LoQ_functional(df = LoQ_F_df, col_lot = "reagent_lot", col_sample = "sample",
col_avg = "avg", col_sd_wl = "sd_wl", target_cv = 10, model = "power",
coeff_start = c(0,0))

}
